Support Elixir 1.15 / OTP 26 and backwards compatibility
OTP 26 adds some control codes to the end of TTY data for controlling the
cursor. It also changed how the terminal buffer is controlled which breaks
with the current implementation.

This adds the needed changes for OTP 26 and also keeps the old TTY control
code in place to use when < OTP 26 is detected for backwards compatibility.

Tests were also modified since Elixir 1.14 and 1.15 both slightly altered
how ANSI data was being adding to the IO