Support Elixir 1.17
The latest Elixir changed how IEx should be started. The current
way we're using is `IEx.start` which is still around in 1.17 but
only arity 0. So if you include any shell_opts, the IEx session would immediately crash. If you don't use shell_opts, then things will work as normal.

This makes the needed adjustments to work with Elixir 1.17 and
support shell_opts, including a change to account for .iex.exs
files in the arguments